server/requirements.txt
Nuitka==2.7.6
ordered-set==4.1.0
pycryptodomex==3.21.0
zstandard==0.23.0
